Centralise all permissions lookup in Link UI and enable
